services/ros-gis-integration/requirements.txt
# FastAPI and GraphQL
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.211.1
python-multipart==0.0.6

# Database connections
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
motor==3.3.2
SQLAlchemy==2.0.23
alembic==1.13.0

# Spatial operations
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
geoalchemy2==0.14.2

# Graph algorithms
networkx==3.2.1

# Data processing
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP clients
httpx==0.25.2
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0